Remove references to sound manager
